nsdocuments/requirements.txt
amqp==2.1.4
appdirs==1.4.0
billiard==3.5.0.2
cassandra-driver==3.7.1
celery==4.0.2
crcmod==1.7
cycler==0.10.0
Cython==0.25.2
decorator==4.0.11
Django==1.10.6
django-angular==0.8.4
django-channels==0.7.0
django-cors-headers==2.0.2
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-extensions==1.7.7
django-filter==1.0.1
django-jet==1.0.5
django-nose==1.4.4
django-python3-ldap==0.9.11
django-redis-sessions==0.5.6
djangorestframework==3.5.4
ipython==5.2.1
ipython-genutils==0.1.0
kombu==4.0.2
ldap3==1.4.0
matplotlib==2.0.0
nanomsg==1.0
netaddr==0.7.19
nose==1.3.7
numexpr==2.6.2
numpy==1.12.0
oauthlib==2.0.1
packaging==16.8
pandas==0.19.2
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.10
protobuf==3.2.0
psycopg2==2.7
ptyprocess==0.5.1
pyasn1==0.2.3
Pygments==2.2.0
pyparsing==2.1.10
PyQt5==5.7.1
python-dateutil==2.6.0
pytz==2016.10
pyzmq==16.0.2
redis==2.10.5
requests==2.13.0
requests-oauthlib==0.8.0
scipy==0.18.1
simplegeneric==0.8.1
sip==4.19
six==1.10.0
sqlparse==0.2.3
tables==3.3.0
traitlets==4.3.1
vine==1.1.3
wcwidth==0.1.7
